pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "subtitle-generator"
version = "0.1.0"
description = "A subtitle generation package using Whisper and M2M100"
authors = [{ name = "Eric Ho", email = "wingsum.developer@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"

dependencies = [
  "click>=7.1",
  "whisper>=1.1.0",
  "transformers",
  "sentencepiece",
  "pytubefix>=9.1.1",
  "webvtt-py",
  "tqdm",
  "ffmpeg-python==0.2.0",
  "numpy",
  "matplotlib",
  "pydub>=0.25.1",
  "python-touch>=0.1",
  "youtube-dl==2021.12.17",
  "behave==1.2.6",
  "pytest>=8.4.0",
  "pytest-mock>=3.14.1",
  "requests>=2.24.1"
]

[project.scripts]
stt = "subtitle_generator.cli_main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["subtitle_generator", "subtitle_generator.*"]
